+++ target/linux/ramips/base-files/lib/upgrade/platform.sh
@@ -120,6 +120,7 @@ platform_check_image() {
 	omega2|\
 	omega2p|\
 	omega2pro|\
+	omega2dash|\
 	omega2lte|\
 	oy-0001|\
 	pbr-d1|\